Update Pullen Census Data Tools.pyt
Fixed the issues keeping the tool from working with the Geoprocessing History.  Enhanced the updateMessages method to address variables not being available across all years and that the output is restricted to a file geodatabase by the code.  Made changes to the runtime messages to be more informative to the user.